requirements.txt
fastjson-db==0.2.1
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.4.4
werkzeug==2.3.7
python-dotenv==1.0.1